# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend API for Cultour Project",
        "title": "Cultour API",
        "contact": {
            "name": "Cultour Development Team"
        },
        "version": "1.0"
    },
    "host": "localhost:8181",
    "paths": {
...
        "/messages": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a list of messages with pagination and filtering",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "List messages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token (without 'Bearer ' prefix)",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of messages to retrieve",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Number of messages to skip",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "\"created_at\"",
                        "description": "Field to sort by",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "\"desc\"",
                        "description": "Sort order (asc/desc)",
                        "name": "sort_order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Messages retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_discussion_models.ResponseMessage"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Failed to list messages",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add a new message to the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "Create a new message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token (without 'Bearer ' prefix)",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "Message Information",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_discussion_models.Message"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Message created successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_discussion_models.ResponseMessage"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid message creation details",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/messages/search": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Search messages by various criteria",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "Search messages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token (without 'Bearer ' prefix)",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Search query (content, etc.)",
                        "name": "query",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of results to retrieve",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Number of results to skip",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "\"created_at\"",
                        "description": "Field to sort by",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "\"desc\"",
                        "description": "Sort order (asc/desc)",
                        "name": "sort_order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Messages found successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_discussion_models.ResponseMessage"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid search parameters",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/messages/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update an existing message's details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "Update a message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token (without 'Bearer ' prefix)",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Message ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Message Update Details",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_discussion_models.Message"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message updated successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_discussion_models.ResponseMessage"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid message update details",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Message not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Remove a message from the system by its unique identifier",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "Delete a message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token (without 'Bearer ' prefix)",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Message ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid message ID",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Message not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_holycann_cultour-backend_internal_discussion_models.Message": {
            "type": "object",
            "required": [
                "content"
            ],
            "properties": {
                "content": {
                    "description": "Message content",
                    "type": "string"
                },
                "created_at": {
                    "description": "Message creation time",
                    "type": "string"
                },
                "id": {
                    "description": "Unique ID for the message",
                    "type": "string"
                },
                "thread_id": {
                    "description": "ID of the thread the message belongs to",
                    "type": "string"
                },
                "type": {
                    "description": "Type of message",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_discussion_models.MessageType"
                        }
                    ]
                },
                "updated_at": {
                    "description": "Message last update time",
                    "type": "string"
                },
                "user_id": {
                    "description": "ID of the user who sent the message",
                    "type": "string"
                }
            }
        },
        "github_com_holycann_cultour-backend_internal_discussion_models.MessageType": {
            "type": "string",
            "enum": [
                "discussion",
                "ai"
            ],
            "x-enum-varnames": [
                "DiscussionMessageType",
                "AIMessageType"
            ]
        },
        "github_com_holycann_cultour-backend_internal_discussion_models.ResponseMessage": {
            "type": "object",
            "required": [
                "content"
            ],
            "properties": {
                "content": {
                    "description": "Message content",
                    "type": "string"
                },
                "created_at": {
                    "description": "Message creation time",
                    "type": "string"
                },
                "id": {
                    "description": "Unique ID for the message",
                    "type": "string"
                },
                "thread_id": {
                    "description": "ID of the thread the message belongs to",
                    "type": "string"
                },
                "type": {
                    "description": "Type of message",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_discussion_models.MessageType"
                        }
                    ]
                },
                "updated_at": {
                    "description": "Message last update time",
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_users_models.User"
                },
                "user_id": {
                    "description": "ID of the user who sent the message",
                    "type": "string"
                }
            }
        },
...
        "github_com_holycann_cultour-backend_internal_response.APIResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Actual response data"
                },
                "error": {
                    "description": "Detailed error information (only populated for error responses)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.ErrorDetails"
                        }
                    ]
                },
                "message": {
                    "description": "Human-readable message",
                    "type": "string"
                },
                "pagination": {
                    "description": "Pagination information (optional)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.Pagination"
                        }
                    ]
                },
                "request_id": {
                    "description": "Unique request identifier for tracing",
                    "type": "string"
                },
                "status": {
                    "description": "Status of the response (success/error)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_holycann_cultour-backend_internal_response.ResponseStatus"
                        }
                    ]
                }
            }
        },
        "github_com_holycann_cultour-backend_internal_response.ErrorDetails": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Machine-readable error code",
                    "type": "string"
                },
                "details": {
                    "description": "Detailed error description",
                    "type": "string"
                }
            }
        },
        "github_com_holycann_cultour-backend_internal_response.Pagination": {
            "type": "object",
            "properties": {
                "has_next_page": {
                    "type": "boolean"
                },
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "github_com_holycann_cultour-backend_internal_response.ResponseStatus": {
            "type": "string",
            "enum": [
                "success",
                "error"
            ],
            "x-enum-varnames": [
                "StatusSuccess",
                "StatusError"
            ]
        },
        "github_com_holycann_cultour-backend_internal_users_models.User": {
            "description": "User account details with authentication and metadata",
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "created_at": {
                    "description": "Timestamp when the user account was created",
                    "type": "string"
                },
                "deleted_at": {
                    "description": "Timestamp of user account soft deletion",
                    "type": "string"
                },
                "email": {
                    "description": "User's email address\n@example \"user@example.com\"",
                    "type": "string",
                    "example": "admin@gmail.com"
                },
                "id": {
                    "description": "Unique identifier for the user\n@example \"user_123\"",
                    "type": "string",
                    "example": "user_123"
                },
                "last_sign_in_at": {
                    "description": "Timestamp of the last user sign-in",
                    "type": "string"
                },
                "phone": {
                    "description": "User's phone number\n@example \"+1234567890\"",
                    "type": "string",
                    "example": "+1234567890"
                },
                "role": {
                    "description": "User's role in the system\n@example \"user\"\n@enums \"user\",\"admin\",\"moderator\"",
                    "type": "string",
                    "enum": [
                        "user",
                        "admin",
                        "moderator"
                    ],
                    "example": "user"
                },
                "updated_at": {
                    "description": "Timestamp of the last user account update",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Insert Your JWT Token. Do NOT include \"Bearer \" prefix. Example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}